+++ lua/convcommit/init.lua
@@ -4,5 +4,6 @@ M.create_version_tag = require("convcommit.version").create_version_tag
 M.create_commit = require("convcommit.create_commit").create_commit
 M.git_add = require("convcommit.git_add").git_add
 M.push = require("convcommit.git_push").push
+M.setup = require("convcommit.setup").setup
 
 return M